Migrate to each type serializer using a module
