+++ rbx_xml/src/serializer.rs
@@ -11,28 +11,7 @@ use rbx_dom_weak::{RbxTree, RbxValue, RbxId};
 
 use crate::{
     reflection::CANONICAL_TO_XML_NAME,
-    types::{
-        float32,
-        float64,
-        int32,
-        int64,
-        serialize_binary_string,
-        serialize_bool,
-        serialize_cframe,
-        serialize_color3,
-        serialize_color3uint8,
-        serialize_content,
-        serialize_enum,
-        serialize_physical_properties,
-        serialize_ref,
-        serialize_string,
-        serialize_udim,
-        serialize_udim2,
-        serialize_vector2,
-        serialize_vector2int16,
-        serialize_vector3,
-        serialize_vector3int16,
-    },
+    types,
 };
 
 pub use xml::writer::XmlEvent as XmlWriteEvent;
@@ -162,26 +141,26 @@ fn serialize_value<W: Write>(
         .unwrap_or(&canonical_name);
 
     match value {
-        RbxValue::BinaryString { value } => serialize_binary_string(writer, xml_name, value),
-        RbxValue::Bool { value } => serialize_bool(writer, xml_name, *value),
-        RbxValue::CFrame { value } => serialize_cframe(writer, xml_name, *value),
-        RbxValue::Color3 { value } => serialize_color3(writer, xml_name, *value),
-        RbxValue::Color3uint8 { value } => serialize_color3uint8(writer, xml_name, *value),
-        RbxValue::Content { value } => serialize_content(writer, xml_name, value),
-        RbxValue::Enum { value } => serialize_enum(writer, xml_name, *value),
-        RbxValue::Float32 { value } => float32::serialize(writer, xml_name, *value),
-        RbxValue::Float64 { value } => float64::serialize(writer, xml_name, *value),
-        RbxValue::Int32 { value } => int32::serialize(writer, xml_name, *value),
-        RbxValue::Int64 { value } => int64::serialize(writer, xml_name, *value),
-        RbxValue::PhysicalProperties { value } => serialize_physical_properties(writer, xml_name, *value),
-        RbxValue::Ref { value } => serialize_ref(writer, xml_name, *value),
-        RbxValue::String { value } => serialize_string(writer, xml_name, value),
-        RbxValue::UDim { value } => serialize_udim(writer, xml_name, *value),
-        RbxValue::UDim2 { value } => serialize_udim2(writer, xml_name, *value),
-        RbxValue::Vector2 { value } => serialize_vector2(writer, xml_name, *value),
-        RbxValue::Vector2int16 { value } => serialize_vector2int16(writer, xml_name, *value),
-        RbxValue::Vector3 { value } => serialize_vector3(writer, xml_name, *value),
-        RbxValue::Vector3int16 { value } => serialize_vector3int16(writer, xml_name, *value),
+        RbxValue::BinaryString { value } => types::binary_string::serialize(writer, xml_name, value),
+        RbxValue::Bool { value } => types::bool::serialize(writer, xml_name, *value),
+        RbxValue::CFrame { value } => types::cframe::serialize(writer, xml_name, *value),
+        RbxValue::Color3 { value } => types::color3::serialize(writer, xml_name, *value),
+        RbxValue::Color3uint8 { value } => types::color3uint8::serialize(writer, xml_name, *value),
+        RbxValue::Content { value } => types::content::serialize(writer, xml_name, value),
+        RbxValue::Enum { value } => types::enumeration::serialize(writer, xml_name, *value),
+        RbxValue::Float32 { value } => types::float32::serialize(writer, xml_name, *value),
+        RbxValue::Float64 { value } => types::float64::serialize(writer, xml_name, *value),
+        RbxValue::Int32 { value } => types::int32::serialize(writer, xml_name, *value),
+        RbxValue::Int64 { value } => types::int64::serialize(writer, xml_name, *value),
+        RbxValue::PhysicalProperties { value } => types::physical_properties::serialize(writer, xml_name, *value),
+        RbxValue::Ref { value } => types::referent::serialize(writer, xml_name, *value),
+        RbxValue::String { value } => types::string::serialize(writer, xml_name, value),
+        RbxValue::UDim { value } => types::udim::serialize(writer, xml_name, *value),
+        RbxValue::UDim2 { value } => types::udim2::serialize(writer, xml_name, *value),
+        RbxValue::Vector2 { value } => types::vector2::serialize(writer, xml_name, *value),
+        RbxValue::Vector2int16 { value } => types::vector2int16::serialize(writer, xml_name, *value),
+        RbxValue::Vector3 { value } => types::vector3::serialize(writer, xml_name, *value),
+        RbxValue::Vector3int16 { value } => types::vector3int16::serialize(writer, xml_name, *value),
 
         unknown => {
             warn!("Property value {:?} cannot be serialized yet", unknown);
