+++ rbx_xml/src/types/mod.rs
@@ -1,25 +1,19 @@
-mod binary_string;
-mod bool;
-mod cframe;
+pub mod binary_string;
+pub mod bool;
+pub mod cframe;
+pub mod content;
+pub mod enumeration;
+pub mod physical_properties;
+pub mod referent;
+
 mod colors;
-mod content;
-mod enumeration;
 mod numbers;
-mod physical_properties;
-mod r#ref;
-mod string;
-mod udim;
+mod strings;
+mod udims;
 mod vectors;
 
-pub use self::binary_string::*;
-pub use self::bool::*;
-pub use self::cframe::*;
-pub use self::colors::*;
-pub use self::content::*;
-pub use self::enumeration::*;
 pub use self::numbers::*;
-pub use self::physical_properties::*;
-pub use self::r#ref::*;
-pub use self::string::*;
-pub use self::udim::*;
+pub use self::strings::*;
+pub use self::udims::*;
 pub use self::vectors::*;
+pub use self::colors::*;
