+++ rbx_xml/src/types/referent.rs
@@ -1,6 +1,6 @@
 use std::io::{Read, Write};
 
-use rbx_dom_weak::RbxValue;
+use rbx_dom_weak::{RbxId, RbxValue};
 
 use crate::{
     deserializer::{DecodeError, EventIterator},
@@ -10,21 +10,26 @@ use crate::{
 pub fn serialize<W: Write>(
     writer: &mut XmlEventWriter<W>,
     name: &str,
-    value: u32,
+    value: Option<RbxId>,
 ) -> Result<(), EncodeError> {
-    writer.write(XmlWriteEvent::start_element("token").attr("name", name))?;
-    writer.write(XmlWriteEvent::characters(&value.to_string()))?;
+    writer.write(XmlWriteEvent::start_element("Ref").attr("name", name))?;
+
+    match value {
+        Some(value) => writer.write(XmlWriteEvent::characters(&value.to_string()))?,
+        None => writer.write(XmlWriteEvent::characters("null"))?,
+    }
+
     writer.write(XmlWriteEvent::end_element())?;
 
     Ok(())
 }
 
 pub fn deserialize<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
-    let value: u32 = reader.read_tag_contents("token")?.parse()?;
+    let _ref_contents = reader.read_tag_contents("Ref")?;
 
-    Ok(RbxValue::Enum {
-        value,
-    })
+    // TODO: Return a different type and use it to figure out the instance's
+    // actual rbx_dom_weak ID, which is separate from Roblox refs.
+    Ok(RbxValue::Ref { value: None })
 }
 
 #[cfg(test)]
@@ -32,22 +37,21 @@ mod test {
     use super::*;
 
     #[test]
-    fn round_trip() {
+    fn round_trip_ref_none() {
         let _ = env_logger::try_init();
 
-        let test_input: u32 = 4654321;
+        let test_input: Option<RbxId> = None;
         let mut buffer = Vec::new();
 
         let mut writer = XmlEventWriter::from_output(&mut buffer);
-        serialize_enum(&mut writer, "foo", test_input).unwrap();
-
+        serialize_ref(&mut writer, "foo", test_input).unwrap();
         println!("{}", std::str::from_utf8(&buffer).unwrap());
 
         let mut reader = EventIterator::from_source(buffer.as_slice());
         reader.next().unwrap().unwrap(); // Eat StartDocument event
-        let value = deserialize_enum(&mut reader).unwrap();
+        let value = deserialize_ref(&mut reader).unwrap();
 
-        assert_eq!(value, RbxValue::Enum {
+        assert_eq!(value, RbxValue::Ref {
             value: test_input,
         });
     }
