+++ rbx_xml/src/types/vectors.rs
@@ -10,106 +10,122 @@ use crate::{
 static VECTOR2_TAGS: [&str; 2] = ["X", "Y"];
 static VECTOR3_TAGS: [&str; 3] = ["X", "Y", "Z"];
 
-pub fn serialize_vector2<W: Write>(
-    writer: &mut XmlEventWriter<W>,
-    name: &str,
-    value: [f32; 2],
-) -> Result<(), EncodeError> {
-    writer.write(XmlWriteEvent::start_element("Vector2").attr("name", name))?;
-    writer.write_tag_array(&value, &VECTOR2_TAGS)?;
-    writer.write(XmlWriteEvent::end_element())?;
+pub mod vector2 {
+    use super::*;
 
-    Ok(())
+    pub fn serialize<W: Write>(
+        writer: &mut XmlEventWriter<W>,
+        name: &str,
+        value: [f32; 2],
+    ) -> Result<(), EncodeError> {
+        writer.write(XmlWriteEvent::start_element("Vector2").attr("name", name))?;
+        writer.write_tag_array(&value, &VECTOR2_TAGS)?;
+        writer.write(XmlWriteEvent::end_element())?;
+
+        Ok(())
+    }
+
+    pub fn deserialize<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
+        reader.expect_start_with_name("Vector2")?;
+
+        let x: f32 = reader.read_tag_contents("X")?.parse()?;
+        let y: f32 = reader.read_tag_contents("Y")?.parse()?;
+
+        reader.expect_end_with_name("Vector2")?;
+
+        Ok(RbxValue::Vector2 {
+            value: [x, y],
+        })
+    }
 }
 
-pub fn deserialize_vector2<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
-    reader.expect_start_with_name("Vector2")?;
+pub mod vector2int16 {
+    use super::*;
 
-    let x: f32 = reader.read_tag_contents("X")?.parse()?;
-    let y: f32 = reader.read_tag_contents("Y")?.parse()?;
+    pub fn serialize<W: Write>(
+        writer: &mut XmlEventWriter<W>,
+        name: &str,
+        value: [i16; 2],
+    ) -> Result<(), EncodeError> {
+        writer.write(XmlWriteEvent::start_element("Vector2int16").attr("name", name))?;
+        writer.write_tag_array(&value, &VECTOR2_TAGS)?;
+        writer.write(XmlWriteEvent::end_element())?;
 
-    reader.expect_end_with_name("Vector2")?;
+        Ok(())
+    }
 
-    Ok(RbxValue::Vector2 {
-        value: [x, y],
-    })
+    pub fn deserialize<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
+        reader.expect_start_with_name("Vector2int16")?;
+
+        let x: i16 = reader.read_tag_contents("X")?.parse()?;
+        let y: i16 = reader.read_tag_contents("Y")?.parse()?;
+
+        reader.expect_end_with_name("Vector2int16")?;
+
+        Ok(RbxValue::Vector2int16 {
+            value: [x, y],
+        })
+    }
 }
 
-pub fn serialize_vector2int16<W: Write>(
-    writer: &mut XmlEventWriter<W>,
-    name: &str,
-    value: [i16; 2],
-) -> Result<(), EncodeError> {
-    writer.write(XmlWriteEvent::start_element("Vector2int16").attr("name", name))?;
-    writer.write_tag_array(&value, &VECTOR2_TAGS)?;
-    writer.write(XmlWriteEvent::end_element())?;
+pub mod vector3 {
+    use super::*;
 
-    Ok(())
+    pub fn serialize<W: Write>(
+        writer: &mut XmlEventWriter<W>,
+        name: &str,
+        value: [f32; 3],
+    ) -> Result<(), EncodeError> {
+        writer.write(XmlWriteEvent::start_element("Vector3").attr("name", name))?;
+        writer.write_tag_array(&value, &VECTOR3_TAGS)?;
+        writer.write(XmlWriteEvent::end_element())?;
+
+        Ok(())
+    }
+
+    pub fn deserialize<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
+        reader.expect_start_with_name("Vector3")?;
+
+        let x: f32 = reader.read_tag_contents("X")?.parse()?;
+        let y: f32 = reader.read_tag_contents("Y")?.parse()?;
+        let z: f32 = reader.read_tag_contents("Z")?.parse()?;
+
+        reader.expect_end_with_name("Vector3")?;
+
+        Ok(RbxValue::Vector3 {
+            value: [x, y, z],
+        })
+    }
 }
 
-pub fn deserialize_vector2int16<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
-    reader.expect_start_with_name("Vector2int16")?;
+pub mod vector3int16 {
+    use super::*;
 
-    let x: i16 = reader.read_tag_contents("X")?.parse()?;
-    let y: i16 = reader.read_tag_contents("Y")?.parse()?;
+    pub fn serialize<W: Write>(
+        writer: &mut XmlEventWriter<W>,
+        name: &str,
+        value: [i16; 3],
+    ) -> Result<(), EncodeError> {
+        writer.write(XmlWriteEvent::start_element("Vector3int16").attr("name", name))?;
+        writer.write_tag_array(&value, &VECTOR3_TAGS)?;
+        writer.write(XmlWriteEvent::end_element())?;
 
-    reader.expect_end_with_name("Vector2int16")?;
+        Ok(())
+    }
 
-    Ok(RbxValue::Vector2int16 {
-        value: [x, y],
-    })
-}
+    pub fn deserialize<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
+        reader.expect_start_with_name("Vector3int16")?;
 
-pub fn serialize_vector3<W: Write>(
-    writer: &mut XmlEventWriter<W>,
-    name: &str,
-    value: [f32; 3],
-) -> Result<(), EncodeError> {
-    writer.write(XmlWriteEvent::start_element("Vector3").attr("name", name))?;
-    writer.write_tag_array(&value, &VECTOR3_TAGS)?;
-    writer.write(XmlWriteEvent::end_element())?;
+        let x: i16 = reader.read_tag_contents("X")?.parse()?;
+        let y: i16 = reader.read_tag_contents("Y")?.parse()?;
+        let z: i16 = reader.read_tag_contents("Z")?.parse()?;
 
-    Ok(())
-}
+        reader.expect_end_with_name("Vector3int16")?;
 
-pub fn deserialize_vector3<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
-    reader.expect_start_with_name("Vector3")?;
-
-    let x: f32 = reader.read_tag_contents("X")?.parse()?;
-    let y: f32 = reader.read_tag_contents("Y")?.parse()?;
-    let z: f32 = reader.read_tag_contents("Z")?.parse()?;
-
-    reader.expect_end_with_name("Vector3")?;
-
-    Ok(RbxValue::Vector3 {
-        value: [x, y, z],
-    })
-}
-
-pub fn serialize_vector3int16<W: Write>(
-    writer: &mut XmlEventWriter<W>,
-    name: &str,
-    value: [i16; 3],
-) -> Result<(), EncodeError> {
-    writer.write(XmlWriteEvent::start_element("Vector3int16").attr("name", name))?;
-    writer.write_tag_array(&value, &VECTOR3_TAGS)?;
-    writer.write(XmlWriteEvent::end_element())?;
-
-    Ok(())
-}
-
-pub fn deserialize_vector3int16<R: Read>(reader: &mut EventIterator<R>) -> Result<RbxValue, DecodeError> {
-    reader.expect_start_with_name("Vector3int16")?;
-
-    let x: i16 = reader.read_tag_contents("X")?.parse()?;
-    let y: i16 = reader.read_tag_contents("Y")?.parse()?;
-    let z: i16 = reader.read_tag_contents("Z")?.parse()?;
-
-    reader.expect_end_with_name("Vector3int16")?;
-
-    Ok(RbxValue::Vector3int16 {
-        value: [x, y, z],
-    })
+        Ok(RbxValue::Vector3int16 {
+            value: [x, y, z],
+        })
+    }
 }
 
 #[cfg(test)]
